remove repairer config file
