--- pkg/datarepair/repairer/config.go
@@ -1,43 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package repairer
-
-import (
-	"context"
-	"time"
-
-	"storj.io/storj/internal/memory"
-	"storj.io/storj/pkg/overlay"
-	ecclient "storj.io/storj/pkg/storage/ec"
-	"storj.io/storj/pkg/storage/segments"
-	"storj.io/storj/pkg/transport"
-)
-
-// Config contains configurable values for repairer
-type Config struct {
-	MaxRepair     int           `help:"maximum segments that can be repaired concurrently" default:"100"`
-	Interval      time.Duration `help:"how frequently checker should audit segments" default:"3600s"`
-	OverlayAddr   string        `help:"Address to contact overlay server through"`
-	PointerDBAddr string        `help:"Address to contact pointerdb server through"`
-	MaxBufferMem  memory.Size   `help:"maximum buffer memory (in bytes) to be allocated for read buffers" default:"4M"`
-}
-
-// GetSegmentRepairer creates a new segment repairer from storeConfig values
-func (c Config) GetSegmentRepairer(ctx context.Context, tc transport.Client) (ss SegmentRepairer, err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	var oc overlay.Client
-	oc, err = overlay.NewClientContext(ctx, tc, c.OverlayAddr)
-	if err != nil {
-		return nil, err
-	}
-	pdbServer := 
-	// pdb, err := pdbclient.NewClientContext(ctx, tc, c.PointerDBAddr, c.APIKey)
-	// if err != nil {
-	// 	return nil, err
-	// }
-
-	ec := ecclient.NewClient(tc, c.MaxBufferMem.Int())
-	return segments.NewSegmentRepairer(oc, ec, pdb), nil
-}
